version 2.0: fully switch to firebase
