+++ app/src/main/AndroidManifest.xml
@@ -16,10 +16,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.CampsiteProject"
-        tools:targetApi="31">
+        tools:targetApi="31"
+        tools:ignore="ExtraText">
 
         <activity
-            android:name=".controller.user.ListCampsiteActivity"
+            android:name=".controller.campsite.ListCampsiteActivity"
             android:exported="true" />
 
         <activity
@@ -37,7 +38,7 @@
             android:name=".test.test1st"
             android:exported="true" />
         <activity
-            android:name=".controller.user.CampsiteDetailActivity"
+            android:name=".controller.campsite.CampsiteDetailActivity"
             android:exported="false" />
 
 <!--        map-->
@@ -51,12 +52,12 @@
             android:name=".MainActivity"
             android:exported="false" />
         <activity
-            android:name=".controller.user.CartActivity"
+            android:name=".controller.campsite.CartActivity"
             android:exported="false" />
         <activity
-            android:name=".controller.user.GearListActivity"
+            android:name=".controller.campsite.GearListActivity"
             android:exported="false" />
-        <activity android:name=".controller.order.OrderHistoryActivity" />
+        <activity android:name=".controller.orders.OrderHistoryActivity" />
             android:exported="true" />
         <activity
             android:name=".controller.user.ChatListActivity"
