+++ app/src/main/res/layout/activity_listcampsite.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".controller.user.ListCampsiteActivity">
+    tools:context=".controller.campsite.ListCampsiteActivity">
 
     <!-- Toolbar -->
     <androidx.appcompat.widget.Toolbar
